removed old flattening approach for new one
