+++ sysopt/solver/canonical_transform.py
@@ -1,24 +1,38 @@
 """Symbol Database for simulation and optimisation."""
 # pylint: disable=invalid-name
-from typing import Optional, List, Union, Callable, Dict, Tuple, Iterable,  NewType
-from dataclasses import dataclass, field
+import copy
+from typing import Optional, List, Union, Callable, Dict, Tuple, Iterable
+from dataclasses import dataclass, field, asdict
 from collections import namedtuple, deque
 
 from sysopt.types import Domain
-from sysopt.block import Block, Composite, Connection, Channel, Port
+from sysopt.block import Block, Composite, Connection, Channel, Port, ComponentBase
 from sysopt.symbolic import (
-    Variable, ExpressionGraph, concatenate, Function, symbolic_vector,
-    get_time_variable, as_function, function_from_graph,
-    restriction_map
+    Variable, ExpressionGraph, concatenate, symbolic_vector,
+    get_time_variable, function_from_graph,
+    restriction_map, as_array, sparse_matrix, Matrix
 )
 
-from sysopt.blocks.block_operations import (
-    to_graph, create_functions_from_block,
-    TableEntry, create_tables_from_blocks,
-    get_projections_for_block
-)
 
 from sysopt import exceptions
+
+@dataclass
+class TableEntry:
+    """Name and index of a block variable."""
+    local_name: str
+    block: str
+    local_index: int
+    global_index: int
+
+    @property
+    def name(self):
+        return f'{self.block}/{self.local_name}'
+
+    def __str__(self):
+        return f'\"{self.name}\", {self.global_index} -> {self.local_index})\n'
+
+    def __repr__(self):
+        return f'{self.global_index} -> ({self.name}, {self.local_index})\n'
 
 @dataclass
 class WireEntry:
@@ -45,20 +59,8 @@ class FlattenedSystem:
 
     @staticmethod
     def from_block(block: Block):
-        # fs = flatten_system(block)
-        #return fs
-        """Creates a new flattened system from the given block"""
-        domain = None
-        *functions, tables = create_functions_from_block(block)
-        for f in functions[1:]:
-            if f:
-                domain = f.domain
-                break
-        assert isinstance(domain, Domain)
-
-        graphs = [to_graph(f) if f else None for f in functions]
-
-        return FlattenedSystem(*graphs, tables, domain=domain)
+        fs = flatten_system(block)
+        return fs
 
 
 @dataclass
@@ -69,12 +71,14 @@ class Quadratures:
     regularisers: List[Variable] = field(default_factory=list)
 
 
-def is_internal(wire: Connection):
+def is_internal(wire: Connection) -> bool:
     src, dest = wire
     return src.block.parent == dest.block.parent
 
 
-def get_ports_and_indices(wire: Connection) -> Iterable[Tuple[Channel, Channel]]:
+def get_ports_and_indices(wire: Connection) \
+    -> Iterable[Tuple[Channel, Channel]]:
+
     src, dest = wire
 
     index_pairing = list(zip(src.indices, dest.indices))
@@ -84,9 +88,91 @@ def get_ports_and_indices(wire: Connection) -> Iterable[Tuple[Channel, Channel]]
     return src_port, dest_port, index_pairing
 
 
+def projection_from_entries(entries: List[TableEntry],
+                            global_dim:int,
+                            local_dim: int) -> Matrix:
+    shape = (local_dim, global_dim)
+    m = sparse_matrix(shape)
+    for entry in entries:
+        row = entry.local_index
+        col = entry.global_index
+        m[row, col] = 1
+    return m
+
+
+def inclusion_from_entries(entries:List[TableEntry],
+                           global_dim: int,
+                           local_dim: int) -> Matrix:
+    proj = projection_from_entries(entries, global_dim, local_dim)
+    return proj.T
+
+
+
+def create_tables_from_block(block: Block) -> Tables:
+
+    tables = {}
+
+    for attribute, attr_object in asdict(block.metadata).items():
+        if attr_object:
+            tables[attribute] = [
+                TableEntry(block=str(block), local_name=name,
+                           local_index=i, global_index=i)
+                for i, name in enumerate(attr_object)
+            ]
+        else:
+            tables[attribute] = []
+
+    return tables
+
+
+def merge_table(table_1: Tables, table_2: Tables) -> Tables:
+    out_table = {}
+    for key in set(table_1.keys()) | set(table_2.keys()):
+        l1 = table_1[key] if key in table_1 else []
+        l2 = table_2[key] if key in table_2 else []
+        out_table[key] = [copy.copy(entry) for entry in l1]
+        offset = len(l1)
+        out_table[key] += [
+            TableEntry(local_name=entry.local_name,
+                       block=entry.block,
+                       local_index=entry.local_index,
+                       global_index=entry.global_index + offset)
+            for entry in l2
+        ]
+    return out_table
+
+
+def create_tables_from_blocks(*blocks):
+
+    tables = [create_tables_from_block(block) for block in blocks]
+
+    base, *rest = tables
+
+    for table in rest:
+        base = merge_table(base, table)
+
+    return base
+
+
+def get_projections_for_block(tables:Tables , block: Block):
+    projectors = {}
+    for attr, local_dim in asdict(block.signature).items():
+        if local_dim == 0:
+            projectors[attr] = None
+            continue
+        entries = sorted([
+            entry for entry in tables[attr] if entry.block == str(block)
+        ], key=lambda entry: entry.local_index)
+        projectors[attr] = projection_from_entries(
+            entries, local_dim=local_dim, global_dim=len(tables[attr])
+        )
+
+    return projectors
+
+
 def find_channel_in_table(table: List[TableEntry],
                           port: Port,
-                          local_index: int):
+                          local_index: int) -> int:
 
     def key(entry: TableEntry):
         return (entry.block == str(port.block)
@@ -115,16 +201,21 @@ def internal_wire_to_table_entries(tables: Tables,
         ))
     return entries
 
-def is_wire_forwarding_inputs(wire):
+
+def is_wire_forwarding_inputs(wire: Connection) -> bool:
     src, dest = wire
     return src.port_type == dest.port_type == 'inputs'
 
 
-def forwarded_input_to_table_entry(tables:Tables, wire: Connection):
+def forwarded_input_to_table_entry(tables:Tables,
+                                   wire: Connection) -> List[TableEntry]:
     src_port, dest_port, indices = get_ports_and_indices(wire)
     entries = []
     for src_i, dest_i in indices:
-        global_index = find_channel_in_table(tables['inputs'], dest_port, dest_i)
+        global_index = find_channel_in_table(tables['inputs'],
+                                             dest_port,
+                                             dest_i)
+
         entries.append(TableEntry(
             local_name=src_port.channel_name(src_i),
             block=str(src_port.block),
@@ -133,7 +224,8 @@ def forwarded_input_to_table_entry(tables:Tables, wire: Connection):
         ))
     return entries
 
-def forwarded_output_to_table_entry(tables: Tables, wire: Connection):
+def forwarded_output_to_table_entry(tables: Tables,
+                                    wire: Connection) -> List[TableEntry]:
     src_port, dest_port, indices = get_ports_and_indices(wire)
     entries = []
     for src_i, dest_i in indices:
@@ -147,7 +239,7 @@ def forwarded_output_to_table_entry(tables: Tables, wire: Connection):
     return entries
 
 
-def create_tables(all_blocks:List[Block]):
+def create_tables(all_blocks:List[Block]) -> Tables:
     tables = create_tables_from_blocks(
         *filter(lambda b: not isinstance(b, Composite), all_blocks)
     )
@@ -176,7 +268,7 @@ def create_tables(all_blocks:List[Block]):
 
     return tables, domain
 
-def tree_to_list(block: Union[Composite, Block]):
+def tree_to_list(block: Union[Composite, Block]) -> List[ComponentBase]:
     fifo = deque()
     fifo.append(block)
     result = []
@@ -192,7 +284,7 @@ def tree_to_list(block: Union[Composite, Block]):
 
 
 
-def create_symbols_from_domain(domain):
+def create_symbols_from_domain(domain) -> Arguments:
     arguments = Arguments(
         t=get_time_variable(),
         x=symbolic_vector('states', domain.states),
@@ -205,7 +297,9 @@ def create_symbols_from_domain(domain):
 
 
 def symbolically_evaluate_initial_conditions(block:Block,
-                                             local_arguments: Arguments):
+                                             local_arguments: Arguments
+                                             ) -> ExpressionGraph:
+
     try:
         x0 = block.initial_state(local_arguments.p)
     except NotImplementedError as ex:
@@ -216,7 +310,9 @@ def symbolically_evaluate_initial_conditions(block:Block,
         raise exceptions.FunctionError(
             block, block.initial_state, ex.args
         ) from ex
-    x0 = concatenate(*x0)
+
+    x0 = as_array(x0)
+
     expected_shape = (block.signature.states,)
     if x0.shape != expected_shape:
         raise exceptions.FunctionError(
@@ -240,7 +336,7 @@ def symbolically_evaluate(block: Block,
         raise exceptions.FunctionError(
             block, func, ex.args
         ) from ex
-    f = concatenate(*f)
+    f = as_array(f)
     if f.shape != (dimension, ):
         raise exceptions.FunctionError(
             block,func,
@@ -252,7 +348,8 @@ def symbolically_evaluate(block: Block,
 
 def symbolically_evaluate_block(tables: Dict,
                                 block: Block,
-                                arguments: Arguments):
+                                arguments: Arguments) -> ExpressionGraph:
+
     proj = get_projections_for_block(tables, block)
     proj_x = proj['states']
     proj_z = proj['constraints']
@@ -284,10 +381,10 @@ def symbolically_evaluate_block(tables: Dict,
     return x0, f, g, h
 
 
-def create_constraints_from_wire_list(
-    wires: List[WireEntry],
-    arguments: Arguments,
-    outputs: ExpressionGraph) -> ExpressionGraph:
+def create_constraints_from_wire_list(wires: List[WireEntry],
+                                      arguments: Arguments,
+                                      outputs: ExpressionGraph
+                                      ) -> ExpressionGraph:
 
     sources, sinks = zip(
         *((wire.source_index, wire.destination_index) for wire in wires)
@@ -296,15 +393,13 @@ def create_constraints_from_wire_list(
                              superset_dimension=len(arguments.u))
     proj_y = restriction_map(indices=list(sources),
                              superset_dimension=outputs.shape[0])
-    u_c = proj_u(arguments.u)
-    y_c = proj_y(outputs)
 
     vector_constraint = proj_u(arguments.u) - proj_y(outputs)
 
     return vector_constraint
 
 
-def flatten_system(root: Composite):
+def flatten_system(root: Composite) -> ExpressionGraph:
     all_blocks = tree_to_list(root)
     leaves = filter(lambda x: not isinstance(x,Composite), all_blocks)
     tables, domain = create_tables(all_blocks)
@@ -348,7 +443,7 @@ def flatten_system(root: Composite):
     proj_y = restriction_map(output_indices, output_map.shape[0])
     outs = proj_y(output_map)
     return FlattenedSystem(
-        initial_conditions=function_from_graph(initial_conditions, symbols.p),
+        initial_conditions=function_from_graph(initial_conditions, [symbols.p]),
         vector_field=function_from_graph(vector_field, symbols),
         output_map=function_from_graph(outs, symbols),
         constraints=function_from_graph(constraints, symbols),
