+++ tests/2_solver_interface/test_integrator.py
@@ -15,7 +15,7 @@
 from sysopt.types import *
 from sysopt.block import Block
 from sysopt.solver import SolverContext
-from sysopt.blocks.block_operations import create_functions_from_block
+
 from sysopt.solver.canonical_transform import FlattenedSystem
 
 class LinearScalarEquation(Block):
@@ -80,24 +80,6 @@ class LinearScalarEquation(Block):
 
 
 class TestSymbolicIntegrator:
-
-    def test_build_functions(self):
-        block = LinearScalarEquation()
-        x0, f, g, h, out_table = create_functions_from_block(block)
-        p = [3, 5]
-        args = [1, [2], [], [], p]
-
-        assert x0(p) == p[1]
-        assert f(*args) == -6
-        assert g(*args) == 2
-        assert f.codomain == 1
-        flat_system = FlattenedSystem.from_block(block)
-        assert flat_system.domain == Domain(1, 1, 0, 0, 2)
-        assert flat_system.vector_field is not None
-        assert flat_system.initial_conditions is not None
-        assert flat_system.output_map is not None
-
-        assert flat_system.vector_field.shape == (1, )
 
     def test_init(self):
         block = LinearScalarEquation()
